requirements.txt
flask==2.2.5
gevent~=22.10.2
colorama~=0.4.4
uuid~=1.30
waitress~=2.1.2
APscheduler==3.10
Werkzeug==2.2.3
setuptools==67.3.2
pycryptodome~=3.17
eventlet~=0.33.3
typing-extensions~=4.5.0
azure-storage-file-share
azure-data-tables==12.4.3
azure-storage-blob==12.14.0
pytest~=7.2.1
pysftp~=0.2.9
cffi~=1.15.1
paramiko~=3.0.0
urllib3~=1.26.14
xmltodict~=0.13.0
flask_login~=0.6.2
